decoder: Refactor `decode_string` to return a `Cow`
The internal `decode_string` function no longer necessarily performs a
new `Vec` allocation when decoding the string. A new `Vec` is allocated
when the string is Huffman encoded, otherwise, simply a borrow of the
corresponding slice of the original buffer is returned.

Clients can decide whether they need to copy the string into an owned
buffer or not.