decoder: Add `decode_with_cb` method; refactor `decode` to use it
The `decode_with_cb` method performs the HPACK decoding of the given
buffer and emits the decoded headers by invoking a given callback
function.

The function receives a `Cow<[u8]>` for both the header name and value.
This means that the implementation of the decoder can be made efficient,
avoiding copies and allocations when the header is indexed in the header
table. It leaves it up to the client to decide how to handle the
headers, i.e. whether to copy to a different buffer or perform some
other processing...

The `decode` method still keeps its old functionality, whereby each
header name and value is automatically copied into an owned `Vec<u8>`,
both for convenience, as well as compatibility with the old API.