decoder: Borrow the header if found in the header table
When decoding, it is possible that the header is already found in the
header table. In this case, the `decode_with_cb` method does not need to
automatically copy the header components into a new buffer, but rather
can simply pass this borrowed representation to the callback and let it
decide whether to do this or not.