decoder: Avoid always allocating memory when decoding string literals
Previously, the decoder would always force a decoded string literal into
an owned buffer and then pass this buffer to the clients.

This commit makes it so that the clients are given the borrowed
representation of the string literal whenever possible, leaving it up to
them to copy it to a buffer or handling it any other way that they
please.

Therefore, at this point, the only allocations that the decoder itself
does are in two scenarios:

- When decoding a Huffman-encoded string literal, a buffer to contain
  the decoded representation is allocated.
- When storing headers in the dynamic header table, each header is
  found in an owned buffer.