encoder: Add a function to encode an int into an io::Write
This makes it possible to simply provide a buffer into which the encoded
integer representation should be placed, without having the function
itself do heap allocations (such as what the `encode_integer` function
does).

For now, `encode_integer` is refactored to be implemented in terms of
the new function, however its utility is questionable and it might be
removed in the future.